verify validity

+++ trading-system-server/src/main/java/com/wsep202/TradingSystem/domain/exception/IllegalMinPriceException.java
@@ -1,7 +1,7 @@
 package com.wsep202.TradingSystem.domain.exception;
 
 public class IllegalMinPriceException extends TradingSystemException{
-    public IllegalMinPriceException(int id, double minPrice){
+    public IllegalMinPriceException(long id, double minPrice){
         super("got invalid minPrice value: "+minPrice+" for discountId: "+id);
     }
 }
